+++ Tetrahedral_remeshing/include/CGAL/tetrahedral_remeshing.h
@@ -18,9 +18,8 @@
 #include <CGAL/Triangulation_3.h>
 #include <CGAL/Mesh_complex_3_in_triangulation_3.h>
 
-#include <CGAL/Tetrahedral_remeshing/Sizing_field.h>
-#include <CGAL/Tetrahedral_remeshing/Uniform_sizing_field.h>
-#include <CGAL/Tetrahedral_remeshing/Complex_cells_selector.h>
+#include <CGAL/Uniform_sizing_field.h>
+#include <CGAL/Tetrahedral_remeshing/internal/property_maps.h>
 
 #include <CGAL/Tetrahedral_remeshing/internal/tetrahedral_adaptive_remeshing_impl.h>
 #include <CGAL/Tetrahedral_remeshing/internal/compute_c3t3_statistics.h>
@@ -29,6 +28,8 @@
 #include <CGAL/boost/graph/named_params_helper.h>
 
 #include <CGAL/property_map.h>
+
+#include <type_traits>
 
 #ifdef CGAL_DUMP_REMESHING_STEPS
 #include <sstream>
@@ -78,13 +79,17 @@ namespace CGAL
 *             and vertex base model of `RemeshingVertexBase_3`.
 * @tparam SLDS is an optional parameter for `Triangulation_3`, that
 *             specifies the type of the spatial lock data structure.
+* @tparam SizingFunction a sizing field functional,
+*             model of `RemeshingSizingField_3`, or type convertible to `double`.
 * @tparam NamedParameters a sequence of \ref bgl_namedparameters "Named Parameters"
 *
 * @param tr the triangulation to be remeshed, of type `Triangulation_3<Traits, TDS, SLDS>`.
 *           `Remeshing_triangulation` is a helper class that satisfies all the requirements
 *           of its template parameters.
-* @param target_edge_length the uniform target edge length. This parameter provides a
+* @param sizing the target edge length. This parameter provides a
 *          mesh density target for the remeshing algorithm.
+*          It can be a number convertible to `double`,
+*          or an instance of a model of `RemeshingSizingField_3`.
 * @param np optional sequence of \ref bgl_namedparameters "Named Parameters"
 *          among the ones listed below
 *
@@ -168,50 +173,27 @@ namespace CGAL
 *  \cgalNamedParamsEnd
 *
 * \sa `CGAL::Tetrahedral_remeshing::Remeshing_triangulation_3`
-*
-* @todo implement non-uniform sizing field instead of uniform target edge length
 */
 template<typename Traits, typename TDS, typename SLDS,
-         typename NamedParameters = parameters::Default_named_parameters>
-void tetrahedral_isotropic_remeshing(
-  CGAL::Triangulation_3<Traits, TDS, SLDS>& tr,
-  const double& target_edge_length,
-  const NamedParameters& np = parameters::default_values())
-{
-  typedef CGAL::Triangulation_3<Traits, TDS, SLDS> Triangulation;
-  tetrahedral_isotropic_remeshing(
-    tr,
-    [target_edge_length](const typename Triangulation::Point& /* p */)
-                        {return target_edge_length;},
-    np);
-}
-
-template<typename Traits, typename TDS, typename SLDS,
-         typename NamedParameters = parameters::Default_named_parameters>
-void tetrahedral_isotropic_remeshing(
-  CGAL::Triangulation_3<Traits, TDS, SLDS>& tr,
-  const float& target_edge_length,
-  const NamedParameters& np = parameters::default_values())
-{
-  typedef CGAL::Triangulation_3<Traits, TDS, SLDS> Triangulation;
-  tetrahedral_isotropic_remeshing(
-    tr,
-    [target_edge_length](const typename Triangulation::Point& /* p */)
-                        {return target_edge_length; },
-    np);
-}
-
-template<typename Traits, typename TDS, typename SLDS,
          typename SizingFunction,
-         typename NamedParameters>
+         typename NamedParameters = parameters::Default_named_parameters
+#ifndef DOXYGEN_RUNNING
+       , typename = std::enable_if_t<
+           std::is_invocable_v<SizingFunction,
+                               typename Traits::Point_3, int, typename TDS::Vertex::Index> >
+#endif
+        >
 void tetrahedral_isotropic_remeshing(
   CGAL::Triangulation_3<Traits, TDS, SLDS>& tr,
   const SizingFunction& sizing,
-  const NamedParameters& np)
+  const NamedParameters& np = parameters::default_values())
 {
-  CGAL_assertion(tr.is_valid());
+  CGAL_expensive_assertion(tr.is_valid());
 
-  typedef CGAL::Triangulation_3<Traits, TDS, SLDS> Tr;
+  using Tr = CGAL::Triangulation_3<Traits, TDS, SLDS>;
+  using Remesher_types
+    = typename Tetrahedral_remeshing::internal::Adaptive_remesher_type_generator
+        <Tr, SizingFunction, NamedParameters>;
 
   using parameters::choose_parameter;
   using parameters::get_parameter;
@@ -219,58 +201,33 @@ void tetrahedral_isotropic_remeshing(
   bool remesh_surfaces = choose_parameter(get_parameter(np, internal_np::remesh_boundaries),
                                           true);
   bool protect = !remesh_surfaces;
-  // bool adaptive = choose_parameter(get_parameter(np, internal_np::adaptive_size),
-  //                              false);
+
   std::size_t max_it = choose_parameter(get_parameter(np, internal_np::number_of_iterations),
                                         1);
   bool smooth_constrained_edges
     = choose_parameter(get_parameter(np, internal_np::smooth_constrained_edges),
                        false);
 
-  typedef typename internal_np::Lookup_named_param_def <
-    internal_np::cell_selector_t,
-    NamedParameters,
-    Tetrahedral_remeshing::Complex_cells_selector<Tr>//default
-  > ::type SelectionFunctor;
-  SelectionFunctor cell_select
+  auto cell_select
     = choose_parameter(get_parameter(np, internal_np::cell_selector),
-                       Tetrahedral_remeshing::Complex_cells_selector<Tr>());
+                                     typename Remesher_types::Default_Selection_functor());
 
-  typedef typename Tr::Vertex_handle Vertex_handle;
-  typedef typename internal_np::Lookup_named_param_def <
-    internal_np::vertex_is_constrained_t,
-    NamedParameters,
-    Constant_property_map<Vertex_handle, bool>//default
-  > ::type VCMap;
-  VCMap vcmap = choose_parameter(get_parameter(np, internal_np::vertex_is_constrained),
-                                 Constant_property_map<Vertex_handle, bool>(false));
+  auto vcmap = choose_parameter(get_parameter(np, internal_np::vertex_is_constrained),
+                                typename Remesher_types::Default_VCMap(false));
 
-  typedef std::pair<typename Tr::Vertex_handle, typename Tr::Vertex_handle> Edge_vv;
-  typedef typename internal_np::Lookup_named_param_def <
-    internal_np::edge_is_constrained_t,
-    NamedParameters,
-    Constant_property_map<Edge_vv, bool>//default
-  > ::type ECMap;
-  ECMap ecmap = choose_parameter(get_parameter(np, internal_np::edge_is_constrained),
-                                 Constant_property_map<Edge_vv, bool>(false));
+  auto ecmap = choose_parameter(get_parameter(np, internal_np::edge_is_constrained),
+                                typename Remesher_types::Default_ECMap(false));
 
-  typedef typename Tr::Facet Facet;
-  typedef typename internal_np::Lookup_named_param_def <
-    internal_np::facet_is_constrained_t,
-    NamedParameters,
-    Constant_property_map<Facet, bool>//default
-  > ::type FCMap;
-  FCMap fcmap = choose_parameter(get_parameter(np, internal_np::facet_is_constrained),
-                                 Constant_property_map<Facet, bool>(false));
+  auto fcmap = choose_parameter(get_parameter(np, internal_np::facet_is_constrained),
+                                typename Remesher_types::Default_FCMap(false));
 
-  typedef typename internal_np::Lookup_named_param_def <
-    internal_np::visitor_t,
-    NamedParameters,
-    Tetrahedral_remeshing::internal::Default_remeshing_visitor
-  > ::type Visitor;
-  Visitor visitor
-    = choose_parameter(get_parameter(np, internal_np::visitor),
-                       Tetrahedral_remeshing::internal::Default_remeshing_visitor());
+  // Advanced and non documented parameters
+  auto visitor = choose_parameter(get_parameter(np, internal_np::visitor),
+                                  typename Remesher_types::Default_Visitor());
+
+  auto nb_extra_iterations
+    = choose_parameter(get_parameter(np, internal_np::nb_flip_smooth_iterations),
+        std::size_t(3));
 
 #ifdef CGAL_TETRAHEDRAL_REMESHING_VERBOSE
   std::cout << "Tetrahedral remeshing ("
@@ -282,8 +239,7 @@ void tetrahedral_isotropic_remeshing(
   std::cout.flush();
 #endif
 
-  typedef Tetrahedral_remeshing::internal::Adaptive_remesher<
-    Tr, SizingFunction, VCMap, ECMap, FCMap, SelectionFunctor, Visitor> Remesher;
+  using Remesher = typename Remesher_types::type;
   Remesher remesher(tr, sizing, protect
                   , vcmap, ecmap, fcmap
                   , smooth_constrained_edges
@@ -299,7 +255,6 @@ void tetrahedral_isotropic_remeshing(
 #endif
 
   // perform remeshing
-  std::size_t nb_extra_iterations = 3;
 #ifdef CGAL_TETRAHEDRAL_REMESHING_NO_EXTRA_ITERATIONS
   nb_extra_iterations = 0;
 #endif
@@ -314,6 +269,34 @@ void tetrahedral_isotropic_remeshing(
     cell_select, "statistics_end.txt");
 #endif
 }
+
+template<typename Traits, typename TDS, typename SLDS, typename FT,
+         typename NamedParameters = parameters::Default_named_parameters
+#ifndef DOXYGEN_RUNNING
+       , typename = std::enable_if_t<
+           !std::is_invocable_v<FT,
+                                typename Traits::Point_3, int, typename TDS::Vertex::Index> >
+#endif
+        >
+void tetrahedral_isotropic_remeshing(
+  CGAL::Triangulation_3<Traits, TDS, SLDS>& tr,
+  const FT target_edge_length,
+  const NamedParameters& np = parameters::default_values())
+{
+  using P = typename Traits::Point_3;
+  using Index = typename TDS::Vertex::Index;
+  using Sizing_FT = typename Traits::FT;
+
+  const Sizing_FT target_ft = static_cast<Sizing_FT>(target_edge_length);
+
+  tetrahedral_isotropic_remeshing(
+    tr,
+    [target_ft]
+      (const P&, const int&, const Index&)
+        { return target_ft; },
+    np);
+}
+
 
 /*!
 * \ingroup PkgTetrahedralRemeshingRef
@@ -379,38 +362,31 @@ convert_to_triangulation_3(
 
   using GT   = typename Tr::Geom_traits;
   using TDS  = typename Tr::Triangulation_data_structure;
+  using Dummy_sizing = CGAL::Uniform_sizing_field<GT>;
 
-  using Vertex_handle = typename Tr::Vertex_handle;
-  using Edge_vv = std::pair<Vertex_handle, Vertex_handle>;
-  using Default_edge_pmap = Constant_property_map<Edge_vv, bool>;
-  using ECMap = typename internal_np::Lookup_named_param_def <
-                internal_np::edge_is_constrained_t,
-                NamedParameters,
-                Default_edge_pmap
-                >::type;
-  using Default_vertex_pmap = Constant_property_map<Vertex_handle, bool>;
-  using VCMap = typename internal_np::Lookup_named_param_def <
-                internal_np::vertex_is_constrained_t,
-                NamedParameters,
-                Default_vertex_pmap
-                >::type;
+  using Remesher_types
+    = typename Tetrahedral_remeshing::internal::Adaptive_remesher_type_generator
+    <Tr, Dummy_sizing, NamedParameters, CornerIndex, CurveIndex>;
 
-  ECMap ecmap = choose_parameter(get_parameter(np, internal_np::edge_is_constrained),
-                                 Default_edge_pmap(false));
-  VCMap vcmap = choose_parameter(get_parameter(np, internal_np::vertex_is_constrained),
-                                 Default_vertex_pmap(false));
-
+  using Default_edge_pmap = typename Remesher_types::Default_ECMap;
+  using ECMap = typename Remesher_types::ECMap;
   if (!std::is_same_v<ECMap, Default_edge_pmap>)
   {
+    ECMap ecmap = choose_parameter(get_parameter(np, internal_np::edge_is_constrained),
+                                   Default_edge_pmap(false));
     for (auto e : c3t3.edges_in_complex())
     {
-      const Edge_vv evv
-        = CGAL::Tetrahedral_remeshing::make_vertex_pair(e);//ordered pair
+      const auto evv = CGAL::Tetrahedral_remeshing::make_vertex_pair(e);//ordered pair
       put(ecmap, evv, true);
     }
   }
+
+  using Default_vertex_pmap = typename Remesher_types::Default_VCMap;
+  using VCMap = typename Remesher_types::VCMap;
   if (!std::is_same_v<VCMap, Default_vertex_pmap>)
   {
+    VCMap vcmap = choose_parameter(get_parameter(np, internal_np::vertex_is_constrained),
+                                   Default_vertex_pmap(false));
     for (auto v : c3t3.vertices_in_complex())
     {
       put(vcmap, v, true);
@@ -432,48 +408,56 @@ convert_to_triangulation_3(
 ////// the empty-sphere property is not respected
 ///////
 template<typename Tr,
-         typename CornerIndex, typename CurveIndex,
-         typename NamedParameters = parameters::Default_named_parameters>
+         typename CornerIndex, typename CurveIndex, typename FT,
+         typename NamedParameters = parameters::Default_named_parameters
+#ifndef DOXYGEN_RUNNING
+       , typename = std::enable_if_t<
+           !std::is_invocable_v<FT,
+             typename Tr::Geom_traits::Point_3, int, typename Tr::Vertex::Index> >
+#endif
+        >
 void tetrahedral_isotropic_remeshing(
   CGAL::Mesh_complex_3_in_triangulation_3<Tr, CornerIndex, CurveIndex>& c3t3,
-  const double& target_edge_length,
+  const FT target_edge_length,
   const NamedParameters& np = parameters::default_values())
 {
-  tetrahedral_isotropic_remeshing(
-    c3t3,
-    [target_edge_length](const typename Tr::Point& /* p */)
-                        {return target_edge_length; },
-    np);
-}
+  using P = typename Tr::Geom_traits::Point_3;
+  using Index = typename Tr::Vertex::Index;
+  using Sizing_FT = typename Tr::Geom_traits::FT;
 
-template<typename Tr,
-         typename CornerIndex, typename CurveIndex,
-         typename NamedParameters = parameters::Default_named_parameters>
-void tetrahedral_isotropic_remeshing(
-  CGAL::Mesh_complex_3_in_triangulation_3<Tr, CornerIndex, CurveIndex>& c3t3,
-  const float& target_edge_length,
-  const NamedParameters& np = parameters::default_values)
-{
+  const Sizing_FT target_ft = static_cast<Sizing_FT>(target_edge_length);
+
   tetrahedral_isotropic_remeshing(
     c3t3,
-    [target_edge_length](const typename Tr::Point& /*p*/)
-                        {return target_edge_length; },
+    [target_ft]
+      (const P&, const int&, const Index&)
+        { return target_ft; },
     np);
 }
 
 template<typename Tr,
          typename CornerIndex, typename CurveIndex,
          typename SizingFunction,
-         typename NamedParameters = parameters::Default_named_parameters>
+         typename NamedParameters = parameters::Default_named_parameters
+#ifndef DOXYGEN_RUNNING
+       , typename = std::enable_if_t<
+           std::is_invocable_v<SizingFunction,
+             typename Tr::Geom_traits::Point_3, int, typename Tr::Vertex::Index> >
+#endif
+        >
 void tetrahedral_isotropic_remeshing(
   CGAL::Mesh_complex_3_in_triangulation_3<Tr, CornerIndex, CurveIndex>& c3t3,
   const SizingFunction& sizing,
   const NamedParameters& np = parameters::default_values())
 {
-  CGAL_assertion(c3t3.triangulation().tds().is_valid());
+  CGAL_expensive_assertion(c3t3.triangulation().tds().is_valid());
 
   using parameters::get_parameter;
   using parameters::choose_parameter;
+
+  using Remesher_types
+    = Tetrahedral_remeshing::internal::Adaptive_remesher_type_generator
+        <Tr, SizingFunction, NamedParameters, CornerIndex, CurveIndex>;
 
   bool remesh_surfaces = choose_parameter(get_parameter(np, internal_np::remesh_boundaries),
                                           true);
@@ -482,52 +466,29 @@ void tetrahedral_isotropic_remeshing(
 
   bool smooth_constrained_edges
     = choose_parameter(get_parameter(np, internal_np::smooth_constrained_edges),
-      false);
+        false);
 
-  typedef typename internal_np::Lookup_named_param_def <
-              internal_np::cell_selector_t,
-              NamedParameters,
-              Tetrahedral_remeshing::Complex_cells_selector<Tr>//default
-              > ::type SelectionFunctor;
-  SelectionFunctor cell_select
+  auto cell_select
     = choose_parameter(get_parameter(np, internal_np::cell_selector),
-                       Tetrahedral_remeshing::Complex_cells_selector<Tr>());
+        typename Remesher_types::Default_Selection_functor());
 
-  typedef typename Tr::Vertex_handle Vertex_handle;
-  typedef typename internal_np::Lookup_named_param_def <
-              internal_np::vertex_is_constrained_t,
-              NamedParameters,
-              Constant_property_map<Vertex_handle, bool>//default
-              > ::type VCMap;
-  VCMap vcmap = choose_parameter(get_parameter(np, internal_np::vertex_is_constrained),
-                                 Constant_property_map<Vertex_handle, bool>(false));
+  auto vcmap = choose_parameter(get_parameter(np, internal_np::vertex_is_constrained),
+                                typename Remesher_types::Default_VCMap(false));
 
-  typedef std::pair<typename Tr::Vertex_handle, typename Tr::Vertex_handle> Edge_vv;
-  typedef typename internal_np::Lookup_named_param_def <
-  internal_np::edge_is_constrained_t,
-              NamedParameters,
-              Constant_property_map<Edge_vv, bool>//default
-              > ::type ECMap;
-  ECMap ecmap = choose_parameter(get_parameter(np, internal_np::edge_is_constrained),
-                                 Constant_property_map<Edge_vv, bool>(false));
+  auto ecmap = choose_parameter(get_parameter(np, internal_np::edge_is_constrained),
+                                typename Remesher_types::Default_ECMap(false));
 
-  typedef typename Tr::Facet Facet;
-  typedef typename internal_np::Lookup_named_param_def <
-  internal_np::facet_is_constrained_t,
-              NamedParameters,
-              Constant_property_map<Facet, bool>//default
-              > ::type FCMap;
-  FCMap fcmap = choose_parameter(get_parameter(np, internal_np::facet_is_constrained),
-                                 Constant_property_map<Facet, bool>(false));
+  auto fcmap = choose_parameter(get_parameter(np, internal_np::facet_is_constrained),
+                                typename Remesher_types::Default_FCMap(false));
 
-  typedef typename internal_np::Lookup_named_param_def <
-              internal_np::visitor_t,
-              NamedParameters,
-              Tetrahedral_remeshing::internal::Default_remeshing_visitor
-              > ::type Visitor;
-  Visitor visitor
+  // Advanced and non documented parameters
+  auto visitor
     = choose_parameter(get_parameter(np, internal_np::visitor),
-                       Tetrahedral_remeshing::internal::Default_remeshing_visitor());
+        typename Remesher_types::Default_Visitor());
+
+  auto nb_extra_iterations
+    = choose_parameter(get_parameter(np, internal_np::nb_flip_smooth_iterations),
+        std::size_t(3));
 
 #ifdef CGAL_TETRAHEDRAL_REMESHING_VERBOSE
   std::cout << "Tetrahedral remeshing ("
@@ -539,8 +500,7 @@ void tetrahedral_isotropic_remeshing(
   std::cout.flush();
 #endif
 
-  typedef Tetrahedral_remeshing::internal::Adaptive_remesher<
-    Tr, SizingFunction, VCMap, ECMap, FCMap, SelectionFunctor, Visitor> Remesher;
+  using Remesher = typename Remesher_types::type;
   Remesher remesher(c3t3, sizing, protect
                     , vcmap, ecmap, fcmap
                     , smooth_constrained_edges
@@ -555,20 +515,19 @@ void tetrahedral_isotropic_remeshing(
 #endif
 
   // perform remeshing
-  std::size_t nb_extra_iterations = 3;
 #ifdef CGAL_TETRAHEDRAL_REMESHING_NO_EXTRA_ITERATIONS
   nb_extra_iterations = 0;
 #endif
 
   remesher.remesh(max_it, nb_extra_iterations);
 
-#ifdef CGAL_TETRAHEDRAL_REMESHING_VERBOSE
+#ifdef CGAL_TETRAHEDRAL_REMESHING_DEBUG
   const double angle_bound = 5.0;
   Tetrahedral_remeshing::debug::dump_cells_with_small_dihedral_angle(
     c3t3.triangulation(),
     angle_bound, cell_select, "bad_cells.mesh");
 #endif
-#ifdef CGAL_TETRAHEDRAL_REMESHING_VERBOSE
+#ifdef CGAL_TETRAHEDRAL_REMESHING_DEBUG
   Tetrahedral_remeshing::internal::compute_statistics(
     c3t3.triangulation(),
     cell_select, "statistics_end.txt");